+++ Examples/AcknowExampleCocoaPods/AcknowExampleCocoaPods/ViewController.swift
@@ -18,14 +18,12 @@ class ViewController: UIViewController {
     }
 
     @IBAction func presentAcknowledgementsSwiftUI(_ sender: UIButton) {
-        if #available(iOS 13.0, *) {
-            guard let path = Bundle.main.path(forResource: "Pods-AcknowExampleCocoaPods-acknowledgements", ofType: "plist") else {
-                return
-            }
-
-            let viewController = UIHostingController(rootView: AcknowNavigationSwiftUIView(plistPath: path))
-            present(viewController, animated: true)
+        guard let url = Bundle.main.url(forResource: "Pods-AcknowExampleCocoaPods-acknowledgements", withExtension: "plist") else {
+            return
         }
+
+        let viewController = UIHostingController(rootView: NavigationView { AcknowListSwiftUIView(plistFileURL: url) })
+        present(viewController, animated: true)
     }
 }