+++ Examples/AcknowExampleManual/AcknowExample/ViewController.swift
@@ -1,7 +1,7 @@
 //
 // ViewController.swift
 //
-// Copyright (c) 2015-2021 Vincent Tourraine (https://www.vtourraine.net)
+// Copyright (c) 2015-2025 Vincent Tourraine (https://www.vtourraine.net)
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal
@@ -35,11 +35,11 @@ class ViewController: UIViewController {
     }
     
     @IBAction func pushAcknowListWithCustomTitlePlainStyle(_ sender: AnyObject) {
-        guard let path = Bundle.main.path(forResource: "Pods-acknowledgements", ofType: "plist") else {
+        guard let url = Bundle.main.url(forResource: "Pods-acknowledgements", withExtension: "plist") else {
             return
         }
 
-        let viewController = AcknowListViewController(plistPath: path, style: .plain)
+        let viewController = AcknowListViewController(plistFileURL: url, style: .plain)
         viewController.title = "Plain Style"
         viewController.headerText = "Visit: https://developer.apple.com"
         navigationController?.pushViewController(viewController, animated: true)
@@ -47,11 +47,11 @@ class ViewController: UIViewController {
     
     @IBAction func pushAcknowListWithCustomTitleInsetGroupedStyle(_ sender: AnyObject) {
         if #available(iOS 13.0, *) {
-            guard let path = Bundle.main.path(forResource: "Pods-acknowledgements", ofType: "plist") else {
+            guard let url = Bundle.main.url(forResource: "Pods-acknowledgements", withExtension: "plist") else {
                 return
             }
 
-            let viewController = AcknowListViewController(plistPath: path, style: .insetGrouped)
+            let viewController = AcknowListViewController(plistFileURL: url, style: .insetGrouped)
             viewController.title = "Grouped Inset Style"
             viewController.headerText = "Visit: https://developer.apple.com"
             navigationController?.pushViewController(viewController, animated: true)
@@ -60,11 +60,14 @@ class ViewController: UIViewController {
 
     @IBAction func pushAcknowListSwiftUI(_ sender: AnyObject) {
         if #available(iOS 13.0, *) {
-            guard let path = Bundle.main.path(forResource: "Pods-acknowledgements", ofType: "plist") else {
+            guard let url = Bundle.main.url(forResource: "Pods-acknowledgements", withExtension: "plist") else {
                 return
             }
 
-            let viewController = UIHostingController(rootView: AcknowListSwiftUIView(plistPath: path))
+            var listViewController = AcknowListSwiftUIView(plistFileURL: url)
+            // Add custom acknowledgement with a repository URL but no text, like a SPM package:
+            listViewController.acknowledgements.append(Acknow(title: "Test", repository: URL(string: "https://developer.apple.com")))
+            let viewController = UIHostingController(rootView: listViewController)
             navigationController?.pushViewController(viewController, animated: true)
         }
     }