+++ Sources/AcknowList/AcknowListSwiftUI.swift
@@ -1,7 +1,7 @@
 //
 // AcknowListSwiftUI.swift
 //
-// Copyright (c) 2015-2021 Vincent Tourraine (https://www.vtourraine.net)
+// Copyright (c) 2015-2025 Vincent Tourraine (https://www.vtourraine.net)
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal
@@ -25,16 +25,13 @@ import SwiftUI
 
 extension Acknow: Identifiable {
     public var id: String {
-        get {
-            title
-        }
+        title
     }
 }
 
 /// View that displays a list of acknowledgements.
-@available(iOS 13.0.0, macOS 10.15.0, watchOS 7.0.0, tvOS 13.0.0, *)
+@available(iOS 13.0.0, macOS 10.15.0, watchOS 7.0.0, tvOS 13.0.0, visionOS 1.0.0, *)
 public struct AcknowListSwiftUIView: View {
-    @Environment(\._isAcknowListSwiftUIViewNavigationTitleHidden) private var isAcknowListSwiftUIViewNavigationTitleHidden
 
     /// The represented array of `Acknow`.
     public var acknowledgements: [Acknow] = []
@@ -45,26 +42,47 @@ public struct AcknowListSwiftUIView: View {
     /// Footer text to be displayed below the list of the acknowledgements.
     public var footerText: String?
 
+    public init() {
+        if let acknowList = AcknowParser.defaultAcknowList() {
+            self.init(acknowList: acknowList)
+        }
+        else {
+            print(
+                "** AcknowList Warning **\n" +
+                "No acknowledgements found.\n" +
+                "Please take a look at https://github.com/vtourraine/AcknowList for instructions.", terminator: "\n")
+            self.init(acknowledgements: [])
+        }
+    }
+
+    public init(acknowList: AcknowList) {
+        acknowledgements = acknowList.acknowledgements
+        headerText = acknowList.headerText
+        footerText = acknowList.footerText
+    }
+
     public init(acknowledgements: [Acknow], headerText: String? = nil, footerText: String? = nil) {
         self.acknowledgements = acknowledgements
         self.headerText = headerText
         self.footerText = footerText
     }
 
-    public init(plistPath: String) {
-        let parser = AcknowParser(plistPath: plistPath)
-        let headerFooter = parser.parseHeaderAndFooter()
-        let header: String?
-        let footer = headerFooter.footer
+    public init(plistFileURL: URL) {
+        guard let data = try? Data(contentsOf: plistFileURL),
+              let acknowList = try? AcknowPodDecoder().decode(from: data) else {
+            self.init(acknowledgements: [], headerText: nil, footerText: nil)
+            return
+        }
 
-        if headerFooter.header != AcknowParser.DefaultHeaderText {
-            header = headerFooter.header
+        let header: String?
+        if acknowList.headerText != AcknowPodDecoder.K.DefaultHeaderText {
+            header = acknowList.headerText
         }
         else {
             header = nil
         }
 
-        self.init(acknowledgements: parser.parseAcknowledgements(), headerText: header, footerText: footer)
+        self.init(acknowledgements: acknowList.acknowledgements, headerText: header, footerText: acknowList.footerText)
     }
 
     struct HeaderFooter: View {
@@ -75,49 +93,87 @@ public struct AcknowListSwiftUIView: View {
                 Text(text)
             }
             else {
-                Text("")
+                EmptyView()
             }
         }
     }
 
     public var body: some View {
-        #if os(iOS) || os(tvOS)
-        if isAcknowListSwiftUIViewNavigationTitleHidden {
-            list
-            .listStyle(GroupedListStyle())
-        } else {
-            list
-            .listStyle(GroupedListStyle())
-            .navigationBarTitle(Text(AcknowLocalization.localizedTitle()))
-        }
-        #else
-        list
-        #endif
-    }
-
-    @ViewBuilder
-    private var list: some View {
+#if os(iOS) || os(tvOS)
         List {
             Section(header: HeaderFooter(text: headerText), footer: HeaderFooter(text: footerText)) {
-                ForEach (acknowledgements) { acknowledgement in
-                    NavigationLink(destination: AcknowSwiftUIView(acknowledgement: acknowledgement)) {
-                        Text(acknowledgement.title)
-                    }
+                ForEach(acknowledgements) { acknowledgement in
+                    AcknowListRowSwiftUIView(acknowledgement: acknowledgement)
                 }
             }
         }
-    }
-
-    @ViewBuilder
-    public func navigationTitle(isHidden: Bool) -> some View {
-        environment(\._isAcknowListSwiftUIViewNavigationTitleHidden, isHidden)
+        .listStyle(GroupedListStyle())
+        .navigationBarTitle(Text(AcknowLocalization.localizedTitle()))
+#else
+        List {
+            Section(header: HeaderFooter(text: headerText), footer: HeaderFooter(text: footerText)) {
+                ForEach(acknowledgements) { acknowledgement in
+                    AcknowListRowSwiftUIView(acknowledgement: acknowledgement)
+                }
+            }
+        }
+#endif
     }
 }
 
-@available(iOS 13.0.0, macOS 10.15.0, watchOS 7.0.0, tvOS 13.0.0, *)
+/// View that displays a row in a list of acknowledgements.
+@available(iOS 13.0.0, macOS 10.15.0, watchOS 7.0.0, tvOS 13.0.0, visionOS 1.0.0, *)
+public struct AcknowListRowSwiftUIView: View {
+
+    /// The represented `Acknow`.
+    public var acknowledgement: Acknow
+
+    /// Indicates if the view controller should try to fetch missing licenses from the GitHub API.
+    public var canFetchLicenseFromGitHub = true
+
+    public var body: some View {
+        if acknowledgement.text != nil || canFetchLicenseFromGitHubAndIsGitHubRepository(acknowledgement) {
+            NavigationLink(destination: AcknowSwiftUIView(acknowledgement: acknowledgement)) {
+                Text(acknowledgement.title)
+            }
+        }
+        else if let repository = acknowledgement.repository,
+                canOpenRepository(for: repository) {
+            Button(action: {
+                repository.openWithDefaultBrowser()
+            }) {
+                Text(acknowledgement.title)
+                    .foregroundColor(.primary)
+            }
+        }
+        else {
+            Text(acknowledgement.title)
+        }
+    }
+
+    private func canOpenRepository(for url: URL) -> Bool {
+        guard let scheme = url.scheme else {
+            return false
+        }
+
+        return scheme == "http" || scheme == "https"
+    }
+
+    private func canFetchLicenseFromGitHubAndIsGitHubRepository(_ acknowledgement: Acknow) -> Bool {
+        if canFetchLicenseFromGitHub,
+           let repository = acknowledgement.repository {
+            return GitHubAPI.isGitHubRepository(repository)
+        }
+        else {
+            return false
+        }
+    }
+}
+
+@available(iOS 13.0.0, macOS 10.15.0, watchOS 7.0.0, tvOS 13.0.0, visionOS 1.0.0, *)
 struct AcknowListSwiftUI_Previews: PreviewProvider {
     static let license = """
-        Copyright (c) 2015-2021 Vincent Tourraine (https://www.vtourraine.net)
+        Copyright (c) 2015-2025 Vincent Tourraine (https://www.vtourraine.net)
 
         Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated documentation files (the "Software"), to deal in the Software without restriction, including without limitation the rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, and to permit persons to whom the Software is furnished to do so, subject to the following conditions:
 
@@ -125,9 +181,11 @@ struct AcknowListSwiftUI_Previews: PreviewProvider {
 
         THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
     """
-    static let acks = [Acknow(title: "Title 1", text: license),
-                       Acknow(title: "Title 2", text: license),
-                       Acknow(title: "Title 3", text: license)]
+    static let acks = [
+        Acknow(title: "Title 1", text: license),
+        Acknow(title: "Title 2", text: license),
+        Acknow(title: "Title 3", text: license),
+    ]
 
     static var previews: some View {
         NavigationView {
@@ -139,7 +197,7 @@ struct AcknowListSwiftUI_Previews: PreviewProvider {
             AcknowListSwiftUIView(acknowledgements: acks, headerText: "Test Header", footerText: "Test Footer")
         }
         .previewDevice(PreviewDevice(rawValue: "iPhone 12"))
-        
+
         NavigationView {
             AcknowListSwiftUIView(acknowledgements: acks, headerText: "Test Header", footerText: "Test Footer")
         }
@@ -154,17 +212,5 @@ struct AcknowListSwiftUI_Previews: PreviewProvider {
             AcknowListSwiftUIView(acknowledgements: acks, headerText: "Test Header", footerText: "Test Footer")
         }
         .previewDevice(PreviewDevice(rawValue: "Mac"))
-    }
-}
-
-// MARK: - Style
-private struct _IsAcknowListSwiftUIViewNavigationTitleHiddenEnvironmentKey: EnvironmentKey {
-    fileprivate static let defaultValue: Bool = false
-}
-
-extension EnvironmentValues {
-    fileprivate var _isAcknowListSwiftUIViewNavigationTitleHidden: Bool {
-        get { self[_IsAcknowListSwiftUIViewNavigationTitleHiddenEnvironmentKey.self] }
-        set { self[_IsAcknowListSwiftUIViewNavigationTitleHiddenEnvironmentKey.self] = newValue }
     }
 }